+++ packages/core/test/unit/WriteApi.test.ts
@@ -156,13 +156,32 @@ describe('WriteApi', () => {
         )
       })
     })
-    it('does not retry write when configured to do so', async () => {
+    it('does not retry write when maxRetries is zero', async () => {
       useSubject({maxRetries: 0, batchSize: 1})
       subject.writeRecord('test value=1')
       await waitForCondition(() => logs.error.length > 0)
       await subject.close().then(() => {
         expect(logs.error).to.length(1)
         expect(logs.warn).is.deep.equal([])
+      })
+    })
+    it('does not retry write when maxRetryTime exceeds', async () => {
+      useSubject({maxRetryTime: 5, batchSize: 1})
+      subject.writeRecord('test value=1')
+      // wait for first attempt to fail
+      await waitForCondition(() => logs.warn.length > 0)
+      // wait for retry attempt to fail on timeout
+      await waitForCondition(() => logs.error.length > 0)
+      await subject.close().then(() => {
+        expect(logs.warn).to.length(1)
+        expect(logs.warn[0][0]).contains(
+          'Write to InfluxDB failed (attempt: 1)'
+        )
+        expect(logs.error).to.length(1)
+        expect(logs.error[0][0]).contains(
+          'Write to InfluxDB failed (attempt: 2)'
+        )
+        expect(logs.error[0][1].toString()).contains('Max retry time exceeded')
       })
     })
     it('does not retry write when writeFailed handler returns a Promise', async () => {
@@ -199,7 +218,7 @@ describe('WriteApi', () => {
       expect(logs.warn).has.length(0)
       expect(count).equals(1)
     })
-    it('implementation uses default notifiers', () => {
+    it('implementation uses expected defaults', () => {
       useSubject({})
       const writeOptions = (subject as any).writeOptions as WriteOptions
       expect(writeOptions.writeFailed).equals(DEFAULT_WriteOptions.writeFailed)
@@ -208,6 +227,7 @@ describe('WriteApi', () => {
       )
       expect(writeOptions.writeSuccess).to.not.throw()
       expect(writeOptions.writeFailed).to.not.throw()
+      expect(writeOptions.randomRetry).equals(true)
     })
   })
   describe('convert point time to line protocol', () => {
@@ -302,6 +322,7 @@ describe('WriteApi', () => {
       useSubject({
         flushInterval: 5,
         maxRetries: 1,
+        randomRetry: false,
         batchSize: 10,
         writeSuccess: writeCounters.writeSuccess,
       })
@@ -369,6 +390,7 @@ describe('WriteApi', () => {
       useSubject({
         flushInterval: 5,
         maxRetries: 1,
+        randomRetry: false,
         batchSize: 10,
         writeSuccess: writeCounters.writeSuccess,
         gzipThreshold: 0,